# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jack Pine Farm Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/batches:
    get:
      operationId: adminListBatches
      tags: [admin, batches]
      summary: List all preorder batches (admin)
      responses:
        "200":
          description: Batch list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PreorderBatch"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateBatch
      tags: [admin, batches]
      summary: Create a preorder batch (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreorderBatch"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/batches/{id}:
    get:
      operationId: adminGetBatch
      tags: [admin, batches]
      summary: Get batch details with linked orders (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreorderBatch"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateBatch
      tags: [admin, batches]
      summary: Update a preorder batch (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatchBody"
      responses:
        "200":
          description: Batch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreorderBatch"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /unsubscribe:
    get:
      operationId: getUnsubscribePreferences
      tags: [notify-me]
      summary: Get unsubscribe preferences by token (public)
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unsubscribe preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsubscribePreferences"
        "404":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: processUnsubscribe
      tags: [notify-me]
      summary: Process unsubscribe request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeBody"
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resubscribe:
    post:
      operationId: processResubscribe
      tags: [notify-me]
      summary: Re-subscribe to notifications (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeBody"
      responses:
        "200":
          description: Re-subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    BatchStatus:
      type: string
      enum:
        - open
        - closed
        - complete
    PreorderBatch:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        status:
          $ref: "#/components/schemas/BatchStatus"
        capacityBirds:
          type: integer
        pricePerLbCentsWhole:
          type: integer
        pricePerLbCentsHalf:
          type: integer
        pricePerLbCentsQuarter:
          type: integer
        notes:
          type: ["string", "null"]
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - name
        - status
        - capacityBirds
        - pricePerLbCentsWhole
        - pricePerLbCentsHalf
        - pricePerLbCentsQuarter
        - notes
        - orderCount
        - createdAt
        - updatedAt
    CreateBatchBody:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        status:
          $ref: "#/components/schemas/BatchStatus"
        capacityBirds:
          type: integer
        pricePerLbCentsWhole:
          type: integer
        pricePerLbCentsHalf:
          type: integer
        pricePerLbCentsQuarter:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - productId
        - name
        - capacityBirds
        - pricePerLbCentsWhole
        - pricePerLbCentsHalf
        - pricePerLbCentsQuarter
    UpdateBatchBody:
      type: object
      properties:
        name:
          type: string
        status:
          $ref: "#/components/schemas/BatchStatus"
        capacityBirds:
          type: integer
        pricePerLbCentsWhole:
          type: integer
        pricePerLbCentsHalf:
          type: integer
        pricePerLbCentsQuarter:
          type: integer
        notes:
          type: ["string", "null"]
...
    UnsubscribePreferences:
      type: object
      properties:
        email:
          type: string
        productName:
          type: ["string", "null"]
        globalUnsubscribe:
          type: boolean
        token:
          type: string
      required:
        - email
        - productName
        - globalUnsubscribe
        - token
    UnsubscribeBody:
      type: object
      properties:
        token:
          type: string
        globalUnsubscribe:
          type: boolean
      required:
        - token