# lib/api-zod/src/generated/types/updateBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */
import type { BatchStatus } from "./batchStatus";

export interface UpdateBatchBody {
  name?: string;
  status?: BatchStatus;
  capacityBirds?: number;
  pricePerLbCentsWhole?: number;
  pricePerLbCentsHalf?: number;
  pricePerLbCentsQuarter?: number;
  /** @nullable */
  notes?: string | null;
}